+++ Programs/Program 3/year.cpp
@@ -164,7 +164,7 @@ void YearIndex::getAllYears()
             } // end while
 
             std::string buildLine = "";
-            for(i = 0; i < myVector.size(); ++i)
+            for(i = 0; i < (int)myVector.size(); ++i)
             {
                 std::stringstream strs;
                 strs << myVector.at(i);
